Move LoadingScreen to lib/components
