--- front-end-mobile/src/features/loading/index.ts
@@ -1,1 +1,0 @@
-export { default as LoadingScreen } from './LoadingScreen';
